Manifest superset + pure validation extraction (W13-MF1)
Ports the source repo's manifest-type superset (nullable render-evidence
structural, DerivationContract, RenderStampLedgerEntry, provenance) —
all additive, proven backward-compatible against the 17 bundled manifests.
Extracts validateManifest/computeFixtureBind/computeFastPathEligible into
manifest-validation.ts, a pure lockstep-core candidate (no fs/provider/zod);
manifest.ts keeps loader I/O and re-exports the surface, so existing
importers are untouched. Zero behavior change; 2,353 unit tests green.
